requirements_deploy.txt
Flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
